refactor(Initializer): replace Initializer class
